fix: add get-peer command
