+++ applications/tari_base_node/src/commands/command/get_block.rs
@@ -7,26 +7,8 @@ use strum::EnumString;
 use tari_common_types::types::HashOutput;
 use tari_utilities::message_format::MessageFormat;
 
-use super::{CommandContext, HandleCommand};
+use super::{CommandContext, HandleCommand, TypeOrHex};
 use crate::commands::{args::FromHex, command_handler::Format};
-
-#[derive(Debug)]
-pub enum HeightOrHex {
-    Height(u64),
-    Hex(FromHex<Vec<u8>>),
-}
-
-impl FromStr for HeightOrHex {
-    type Err = Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        if let Ok(hex) = FromHex::<Vec<u8>>::from_str(s) {
-            Ok(Self::Hex(hex))
-        } else {
-            u64::from_str(s).map(Self::Height).map_err(Error::from)
-        }
-    }
-}
 
 /// Display a block by height or hash
 #[derive(Debug, Parser)]
@@ -34,7 +16,7 @@ pub struct Args {
     /// The height or hash of the block to fetch
     /// from the main chain. The genesis block
     /// has height zero.
-    value: HeightOrHex,
+    value: TypeOrHex<u64>,
     /// Supported options are 'json' and 'text'. 'text' is the default if omitted.
     #[clap(default_value_t)]
     format: Format,
@@ -45,8 +27,8 @@ impl HandleCommand<Args> for CommandContext {
     async fn handle_command(&mut self, args: Args) -> Result<(), Error> {
         let format = args.format;
         match args.value {
-            HeightOrHex::Height(height) => self.get_block(height, format).await,
-            HeightOrHex::Hex(hex) => self.get_block_by_hash(hex.0, format).await,
+            TypeOrHex::Type(value) => self.get_block(value, format).await,
+            TypeOrHex::Hex(hex) => self.get_block_by_hash(hex.0, format).await,
         }
     }
 }
