+++ applications/tari_base_node/src/commands/parser.rs
@@ -38,7 +38,6 @@ use strum_macros::{Display, EnumIter, EnumString};
 #[strum(serialize_all = "kebab_case")]
 pub enum BaseNodeCommand {
     Help,
-    GetPeer,
     Quit,
     Exit,
 }
